# contract/openapi.json
{
  "components": {
    "schemas": {
      "AgentCreate": {
        "properties": {
          "config": {
            "discriminator": {
              "mapping": {
                "phone": "#/components/schemas/PhoneConfig",
                "sip": "#/components/schemas/SipConfig",
                "web": "#/components/schemas/WebConfig"
              },
              "propertyName": "transport"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/WebConfig"
              },
              {
                "$ref": "#/components/schemas/SipConfig"
              },
              {
                "$ref": "#/components/schemas/PhoneConfig"
              }
            ],
            "title": "Config"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "maxConcurrency": {
            "default": 1,
            "title": "Maxconcurrency",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "config"
        ],
        "title": "AgentCreate",
        "type": "object"
      },
      "AgentDetail": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "lastSeenAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastseenat"
          },
          "maxConcurrency": {
            "title": "Maxconcurrency",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "projectId": {
            "title": "Projectid",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "transport": {
            "enum": [
              "web",
              "sip",
              "phone"
            ],
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "projectId",
          "transport",
          "config",
          "maxConcurrency"
        ],
        "title": "AgentDetail",
        "type": "object"
      },
...
      "AgentUpdate": {
        "properties": {
          "config": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "phone": "#/components/schemas/PhoneConfig",
                    "sip": "#/components/schemas/SipConfig",
                    "web": "#/components/schemas/WebConfig"
                  },
                  "propertyName": "transport"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/WebConfig"
                  },
                  {
                    "$ref": "#/components/schemas/SipConfig"
                  },
                  {
                    "$ref": "#/components/schemas/PhoneConfig"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Config"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "maxConcurrency": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maxconcurrency"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "AgentUpdate",
        "type": "object"
      },
...
      "ConnectedAgent": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "meta": {
            "title": "Meta",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "Healthy",
              "Degraded",
              "Offline"
            ],
            "title": "Status",
            "type": "string"
          },
          "tag": {
            "enum": [
              "RTC",
              "SIP",
              "TEL"
            ],
            "title": "Tag",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "meta",
          "tag",
          "status"
        ],
        "title": "ConnectedAgent",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PhoneConfig": {
        "properties": {
          "phoneNumber": {
            "title": "Phonenumber",
            "type": "string"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "transport": {
            "const": "phone",
            "default": "phone",
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "phoneNumber"
        ],
        "title": "PhoneConfig",
        "type": "object"
      },
...
      "SipConfig": {
        "properties": {
          "password": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "sipUri": {
            "title": "Sipuri",
            "type": "string"
          },
          "transport": {
            "const": "sip",
            "default": "sip",
            "title": "Transport",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "sipUri"
        ],
        "title": "SipConfig",
        "type": "object"
      },
...
      "TestConnectionResult": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "rttMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rttms"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "detail"
        ],
        "title": "TestConnectionResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WebConfig": {
        "properties": {
          "roomUrl": {
            "title": "Roomurl",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          },
          "transport": {
            "const": "web",
            "default": "web",
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "roomUrl",
          "token"
        ],
        "title": "WebConfig",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "cadence-brain",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/agents": {
      "get": {
        "operationId": "list_agents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectedAgent"
                  },
                  "title": "Response List Agents",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "create_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Agent",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/v1/agents/test-connection": {
      "post": {
        "operationId": "test_connection_pre_save",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "discriminator": {
                  "mapping": {
                    "phone": "#/components/schemas/PhoneConfig",
                    "sip": "#/components/schemas/SipConfig",
                    "web": "#/components/schemas/WebConfig"
                  },
                  "propertyName": "transport"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/WebConfig"
                  },
                  {
                    "$ref": "#/components/schemas/SipConfig"
                  },
                  {
                    "$ref": "#/components/schemas/PhoneConfig"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection Pre Save",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/{agent_id}": {
      "delete": {
        "operationId": "delete_agent",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Agent",
        "tags": [
          "agents"
        ]
      },
      "get": {
        "operationId": "get_agent",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Agent",
        "tags": [
          "agents"
        ]
      },
      "patch": {
        "operationId": "update_agent",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/{agent_id}/test-connection": {
      "post": {
        "operationId": "test_connection_post_save",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection Post Save",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}